rename : 폴더 이름 Camel -> Kebab
