+++ mcnc_2024/src/components/create-survey/create-survey-item/type-component/ObjComponent.vue
